chapter_08_prompt_chaining/practical_example_01/requirements.txt
langchain ~= 0.3.4
langchain-community ~= 0.3.3
langchain-core ~= 0.3.13
langchain-openai ~= 0.2.3
pydantic ~= 2.9.2
python-dotenv ~= 1.0.1